moved routes into config
