+++ config/routes.js
@@ -1,7 +1,13 @@
+const {MongoClient} = require("mongodb");
 
 function routes(app){
-    app.get('/', )
-
+    app.get('/', async (request, response) => {
+        const connection = await MongoClient.connect(mongoConnection);
+        const db = connection.db('canineCompass')
+        const collection = db.collection('dogWalks')
+        const data = await collection.find({}).toArray()
+        response.json(data)
+    })
 }
 
 module.exports = routes
