+++ index.js
@@ -13,14 +13,6 @@ app.set('view engine', 'handlebars')
 app.use(express.json());
 app.use(express.static('public'));
 
-app.get('/', async (request, response) => {
-    const connection = await MongoClient.connect(mongoConnection);
-    const db = connection.db('canineCompass')
-    const collection = db.collection('dogWalks')
-    const data = await collection.find({}).toArray()
-    response.json(data)
-})
-
 router(app)
 
 app.listen(port)
